+++ src/modules/Classes/Http/Controllers/GradeController.php
@@ -17,6 +17,8 @@ class GradeController extends BaseController
 
 	public function getGradeClassDelete($gradeId, $classId)
 	{
+		$this->authorize('edit_grade');
+		
 		$this->classRepository->deleteClass($classId, $gradeId);
 
 		return $this->printJson(true, [], trans('classes::class.class_deleted'));
@@ -24,6 +26,8 @@ class GradeController extends BaseController
 
 	public function postGradeClassEdit(UpdateClassRequest $request, $gradeId, $classId)
 	{
+		$this->authorize('edit_grade');
+		
 		$class = $this->classRepository->updateClass($request->all(), $classId, $gradeId);
 
 		return $this->printPartial(view('classes::partials.class', [
@@ -34,6 +38,8 @@ class GradeController extends BaseController
 
 	public function getGradeClassEdit($gradeId, $classId)
 	{
+		$this->authorize('edit_grade');
+		
 		return $this->printModal(view('classes::modals.edit_class', [
 				'class' => $this->classRepository->findClass($classId, $gradeId),
 				'grade' => $this->classRepository->findGrade($gradeId),
@@ -43,6 +49,8 @@ class GradeController extends BaseController
 
 	public function postGradeClassNew(CreateClassRequest $request, $gradeId)
 	{
+		$this->authorize('edit_grade');
+		
 		$class = $this->classRepository->createClass($request->all(), $gradeId);
 
 		return $this->printPartial(view('classes::partials.class', [
@@ -53,6 +61,8 @@ class GradeController extends BaseController
 
 	public function getGradeClassNew($gradeId)
 	{
+		$this->authorize('edit_grade');
+		
 		return $this->printModal(view('classes::modals.edit_class', [
 				'class' => null,
 				'grade' => $this->classRepository->findGrade($gradeId),
@@ -62,21 +72,29 @@ class GradeController extends BaseController
 
 	public function getGradeDetailsView($gradeId)
 	{
+		$this->authorize('view_grade');
+		
 		return view('classes::view_grade_details', ['grade' => $this->classRepository->findGrade($gradeId)]);
 	}
 
 	public function getGradeClassesView($gradeId)
 	{
+		$this->authorize('view_grade');
+		
 		return view('classes::view_grade_classes', ['grade' => $this->classRepository->findGrade($gradeId)]);
 	}
 
 	public function getGradeClassesEdit($gradeId)
 	{
+		$this->authorize('edit_grade');
+		
 		return view('classes::edit_grade_classes', ['grade' => $this->classRepository->findGrade($gradeId)]);
 	}
 
 	public function postGradeDetailsEdit(UpdateGradeRequest $request, $gradeId)
 	{
+		$this->authorize('edit_grade');
+		
 		$this->classRepository->updateGrade($request->all(), $gradeId);
 
 		return $this->printJson(true, [], trans('classes::grade.grade_updated'));
@@ -84,6 +102,8 @@ class GradeController extends BaseController
 
 	public function getGradeDetailsEdit($gradeId)
 	{
+		$this->authorize('edit_grade');
+		
 		return view('classes::edit_grade_details', [
 				'grade' => $this->classRepository->findGrade($gradeId),
 				'grade_form_validator' => $this->jsValidator(UpdateGradeRequest::class)
@@ -92,6 +112,8 @@ class GradeController extends BaseController
 
 	public function postGradeNew(CreateGradeRequest $request)
 	{
+		$this->authorize('create_grade');
+		
 		$grade = $this->classRepository->createGrade($request->all());
 
 		return $this->printRedirect(route('grade.classes.edit', $grade->id));
@@ -99,6 +121,8 @@ class GradeController extends BaseController
 
 	public function getGradeNew()
 	{
+		$this->authorize('create_grade');
+		
 		return view('classes::edit_grade_details', [
 				'grade' => null,
 				'grade_form_validator' => $this->jsValidator(CreateGradeRequest::class)
@@ -107,6 +131,8 @@ class GradeController extends BaseController
 
 	public function getGradeList(Request $request)
 	{
+		$this->authorize('view_grade');
+		
 		return view('classes::grades_list', [
 				'grades' => $this->classRepository->getGrades()->paginate(config('collejo.pagination.perpage'))
 			]);
@@ -114,6 +140,8 @@ class GradeController extends BaseController
 
 	public function getGradeClasses(Request $request)
 	{
+		$this->authorize('view_batch');
+		
 		return $this->printJson(true, $this->classRepository->findGrade($request::get('grade_id'))->classes->pluck('name', 'id'));
 	}
